requirements.txt
fastapi[standard]
pydantic
python-dotenv
python-jose